# api/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Use your Android phone to send and receive SMS messages via a simple programmable API with end-to-end encryption.",
        "title": "httpSMS API Reference",
        "contact": {
            "name": "support@httpsms.com",
            "email": "support@httpsms.com"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://raw.githubusercontent.com/NdoleStudio/http-sms-manager/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "api.httpsms.com",
    "basePath": "/v1",
    "paths": {
...
        "/phones": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get list of phones which a user has registered on the http sms application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phones"
                ],
                "summary": "Get phones of a user",
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "number of heartbeats to skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter phones containing query",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "maximum": 20,
                        "minimum": 1,
                        "type": "integer",
                        "description": "number of phones to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PhonesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.Unauthorized"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.UnprocessableEntity"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates properties of a user's phone. If the phone with this number does not exist, a new one will be created. Think of this method like an 'upsert'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phones"
                ],
                "summary": "Upsert Phone",
                "parameters": [
                    {
                        "description": "Payload of new phone number.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.PhoneUpsert"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PhoneResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.Unauthorized"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.UnprocessableEntity"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerError"
                        }
                    }
                }
            }
        },
        "/phones/fcm-token": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the FCM token of a phone. If the phone with this number does not exist, a new one will be created. Think of this method like an 'upsert'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phones"
                ],
                "summary": "Upserts the FCM token of a phone",
                "parameters": [
                    {
                        "description": "Payload of new FCM token.",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.PhoneFCMToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PhoneResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.Unauthorized"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.UnprocessableEntity"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerError"
                        }
                    }
                }
            }
        },
        "/phones/{phoneID}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a phone that has been sored in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Phones"
                ],
                "summary": "Delete Phone",
                "parameters": [
                    {
                        "type": "string",
                        "default": "32343a19-da5e-4b1b-a767-3298a73703ca",
                        "description": "ID of the phone",
                        "name": "phoneID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/responses.NoContent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequest"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.Unauthorized"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.UnprocessableEntity"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.Phone": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:02.302718+03:00"
                },
                "fcm_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzd....."
                },
                "id": {
                    "type": "string",
                    "example": "32343a19-da5e-4b1b-a767-3298a73703cb"
                },
                "max_send_attempts": {
                    "description": "MaxSendAttempts determines how many times to retry sending an SMS message",
                    "type": "integer",
                    "example": 2
                },
                "message_expiration_seconds": {
                    "description": "MessageExpirationSeconds is the duration in seconds after sending a message when it is considered to be expired.",
                    "type": "integer"
                },
                "messages_per_minute": {
                    "type": "integer",
                    "example": 1
                },
                "missed_call_auto_reply": {
                    "type": "string",
                    "example": "This phone cannot receive calls. Please send an SMS instead."
                },
                "phone_number": {
                    "type": "string",
                    "example": "+18005550199"
                },
                "sim": {
                    "$ref": "#/definitions/entities.SIM"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2022-06-05T14:26:10.303278+03:00"
                },
                "user_id": {
                    "type": "string",
                    "example": "WB7DRDWrJZRGbYrv2CKGkqbzvqdC"
                }
            }
        },
...
        "entities.SIM": {
            "type": "string",
            "enum": [
                "SIM1",
                "SIM2"
            ],
            "x-enum-varnames": [
                "SIM1",
                "SIM2"
            ]
        },
...
        "requests.PhoneFCMToken": {
            "type": "object",
            "properties": {
                "fcm_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzd....."
                },
                "phone_number": {
                    "type": "string",
                    "example": "[+18005550199]"
                },
                "sim": {
                    "description": "SIM is the SIM slot of the phone in case the phone has more than 1 SIM slot",
                    "type": "string",
                    "example": "SIM1"
                }
            }
        },
        "requests.PhoneUpsert": {
            "type": "object",
            "properties": {
                "fcm_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzd....."
                },
                "max_send_attempts": {
                    "description": "MaxSendAttempts is the number of attempts when sending an SMS message to handle the case where the phone is offline.",
                    "type": "integer",
                    "example": 2
                },
                "message_expiration_seconds": {
                    "description": "MessageExpirationSeconds is the duration in seconds after sending a message when it is considered to be expired.",
                    "type": "integer",
                    "example": 12345
                },
                "messages_per_minute": {
                    "type": "integer",
                    "example": 1
                },
                "missed_call_auto_reply": {
                    "type": "string",
                    "example": "e.g. This phone cannot receive calls. Please send an SMS instead."
                },
                "phone_number": {
                    "type": "string",
                    "example": "+18005550199"
                },
                "sim": {
                    "description": "SIM is the SIM slot of the phone in case the phone has more than 1 SIM slot",
                    "type": "string",
                    "example": "SIM1"
                }
            }
        },
...
        "responses.BadRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "The request body is not a valid JSON string"
                },
                "message": {
                    "type": "string",
                    "example": "The request isn't properly formed"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "responses.InternalServerError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "We ran into an internal error while handling the request."
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "responses.NoContent": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "action performed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "responses.PhoneResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entities.Phone"
                },
                "message": {
                    "type": "string",
                    "example": "Request handled successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "responses.PhonesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Phone"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Request handled successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "responses.Unauthorized": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "Make sure your API key is set in the [X-API-Key] header in the request"
                },
                "message": {
                    "type": "string",
                    "example": "You are not authorized to carry out this request."
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "responses.UnprocessableEntity": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "validation errors while handling request"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-Key",
            "in": "header"
        }
    }
}